Removed TransactionManager.current() from Transaction.pgAdvisoryXactLock() Moved the test files back to their original folder
