+++ src/test/kotlin/com/cultureamp/eventsourcing/PgTestConfig.kt
@@ -1,11 +1,11 @@
-package com.cultureamp.pg
+package com.cultureamp.eventsourcing
 
 import org.jetbrains.exposed.sql.Database
 
 // H2 is the default databse. To use this, you need to input your database details
 object PgTestConfig {
     val testerName = ""
-    val jdbcUrl = "jdbc:postgresql://localhost:5432/${testerName}"
+    val jdbcUrl = "jdbc:postgresql://localhost:5432/$testerName"
     val driver =  "org.postgresql.Driver"
     val user = testerName
